Struct qvs20::qvs20_table_schema_mod::TableSchema

pub struct TableSchema {
    pub table_name: String,
    pub table_description: String,
    pub data_types: Vec<DataType>,
    pub sub_table_schemas: Vec<Option<TableSchema>>,
    pub additional_properties: Vec<String>,
    pub column_names: Vec<String>,
    pub active_row: usize,
    pub active_column: usize,
    pub row_delimiter: u8,
    sub_table_row_delimiter: u8,
}

Fields
table_name: String

Schema 1st row - table name

table_description: Stringdata_types: Vec<DataType>

Schema 2nd row - data types

sub_table_schemas: Vec<Option<TableSchema>>

Schema 3rd row - SubTable schema

additional_properties: Vec<String>

Schema 4th row - additional properties as strings

column_names: Vec<String>

Schema 5th row - column names, the last schema row

active_row: usize

index 0 for first row

active_column: usize

active column

row_delimiter: u8

row_delimiter

sub_table_row_delimiter: u8

sub table row delimiter

Methods

impl TableSchema

pub fn new_simple_strings(count_of_column: usize) -> TableSchema

a simple table with all Strings and automatic column_names

pub fn schema_from_qvs20_str(input: &str) -> Result<TableSchema, Qvs20Error>

create a Schema object in memory from a qvs20 format string in the string is only the schema separated from the data

pub fn read_schema(
&mut self,
rdr: &mut ReaderForQvs20
) -> Result<(), Qvs20Error>

create a table schema in memory from a qvs20 format string

fn active_row_str(&self) -> String

fn read_row_schema_1st_row_table_name_and_description(
&mut self,
rdr: &mut ReaderForQvs20
) -> Result<(), Qvs20Error>

Schema 1st row: table name, description and row_delimiter are mandatory

fn read_row_schema_2nd_row_data_types(
&mut self,
rdr: &mut ReaderForQvs20
) -> Result<(), Qvs20Error>

Schema 2nd row: data types

fn read_row_schema_3rd_row_sub_table_schemas(
&mut self,
rdr: &mut ReaderForQvs20
) -> Result<(), Qvs20Error>

Schema 3rd row - sub tables schema

fn read_row_schema_4th_row_additional_properties(
&mut self,
rdr: &mut ReaderForQvs20
) -> Result<(), Qvs20Error>

Schema 4th row - additional properties

fn read_row_schema_5th_row_column_names(
&mut self,
rdr: &mut ReaderForQvs20
) -> Result<(), Qvs20Error>

Schema 5th row - column names

fn read_field_in_row(
&mut self,
rdr: &mut ReaderForQvs20
) -> Option<Result<ValueForSchema, Qvs20Error>>

read field in row, returns ValueForSchema: variants String or SubTable Schema None means end of the row

pub fn write_schema(&self) -> String

pub fn write_schema_to_writer(
&self,
wrt: &mut WriterForQvs20,
schema_only: bool
)

write to writer