Struct qvs20::WriterForQvs20

pub struct WriterForQvs20<'a> {
    output: &'a mut String,
    schema: &'a TableSchema,
    column: usize,
    row_delimiter: char,
}

Fields
output: &'a mut Stringschema: &'a TableSchemacolumn: usizerow_delimiter: char
Methods

impl<'a> WriterForQvs20<'a>

pub fn new(output: &'a mut String, schema: &'a TableSchema) -> Self

pub fn write_schema(&mut self)

pub fn write_table_rows(&mut self, table_rows: &'a TableRows)

pub fn write_table(&mut self, table_rows: &'a TableRows)

pub fn write_delimiter(&mut self)

pub fn write_string(&mut self, data: &str)

write a field of type String

pub fn write_integer(&mut self, data: i64)

write a field of type integer

pub fn write_decimal(&mut self, data: Decimal)

write a field of type decimal

pub fn write_float(&mut self, data: f64)

write a field of type float

pub fn write_sub_table_schema(&mut self, schema: &TableSchema)

write a sub table schema write a field of type String

pub fn push_escaped_qvs20_str(&mut self, text: &str)

escape the 6 special characters \, [, ], \n, \r, \t all this characters are ascii7 therefore I can use a faster vector of bytes and not a string less escaping needed, faster the performance the parameter escaped is allocated before this fn call